Divided the Game page to api calls, hooks and view
