+++ frontend/src/api/ShelvesConfig.js
@@ -1,6 +1,6 @@
 import axios from "axios";
 
-export const fetchRowConfigs = async () => {
+export const fetchShelvesConfigs = async () => {
     const res = await axios.get("http://localhost:8000/api/games/row-config");
     console.log(res.data);
     return res.data;
